refactor: add LogServiceInterface to improve testability
